backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pillow==10.4.0
python-multipart==0.0.9
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
